Added newest pbsolver of yoav. Still very buggy apparently, but suspecting it is due to a compilation issue (all old versions worked)

+++ lib/pbsolver/pbbase/h/BaseSearchState.h
@@ -1,17 +1,15 @@
 #ifndef BASESEARCHSTATE_H_
 #define BASESEARCHSTATE_H_
 
-#include "Global.h"
-
 namespace MiniSatPP {
 struct BaseSearchState {
     BaseSearchState* parent;
     int index;
 	int lastRelevent;
 	unsigned int baseMul;
-	uint64 cost;
-	uint64 hCost;
-	uint64 carryins;
+	unsigned long long cost;
+	unsigned long long hCost;
+	unsigned long long carryins;
 	BaseSearchState(BaseSearchState* parent_, int lastRelevent_ ,unsigned int baseMul_,
 					unsigned long cost_,unsigned long hCost_,
 					unsigned long carryins_) :
