+++ lib/pbsolver/pbbase/h/BaseSearchStateRelativeComp.h
@@ -4,7 +4,6 @@
 #include <iostream>
 #include <math.h>
 #include <stdio.h>
-#include "Global.h"
 
 namespace MiniSatPP {
 #define gamma  0.134380000001 // 0.129030517578
@@ -16,22 +15,22 @@ struct BaseSearchStateRelativeComp {
     int index;
 	int lastRelevent;
 	unsigned int baseMul;
-	uint64 cnfSize;
-	uint64 carryCost;
-	uint64 hCnfSize;
-	uint64 currentChain;
-	uint64 carryins;
+	unsigned long long cnfSize;
+	unsigned long long carryCost;
+	unsigned long long hCnfSize;
+	unsigned long long currentChain;
+	unsigned long long carryins;
 	BaseSearchStateRelativeComp(BaseSearchStateRelativeComp* parent_, int lastRelevent_ ,unsigned int baseMul_,
-					uint64 cnfSize_, uint64 carryCost_,
-					uint64 hCnfSize_,uint64 currentChain_,
-					uint64 carryins_) :
+					unsigned long long cnfSize_, unsigned long long carryCost_,
+					unsigned long long hCnfSize_,unsigned long long currentChain_,
+					unsigned long long carryins_) : 
 					parent(parent_), index(0),lastRelevent(lastRelevent_) ,baseMul(baseMul_),
 					cnfSize(cnfSize_),  carryCost(carryCost_),
 					hCnfSize(hCnfSize_),currentChain(currentChain_),
 					carryins(carryins_) {};
 };
 
-inline double ratio(uint64 x,uint64 y) {
+inline double ratio(unsigned long long x,unsigned long long y) {
 	if (x==y)       return 0.5;
 	else if (x < y) {
 		if (y==0) printf("zero devsion");
@@ -44,7 +43,7 @@ inline double ratio(uint64 x,uint64 y) {
 }
        
 //returns true iff (a,b) is better then (c,d)
-inline bool relComp(uint64 a,uint64 b, uint64 c, uint64 d) {
+inline bool relComp(unsigned long long a,unsigned long long b, unsigned long long c, unsigned long long d) {
 	//std::cout<<"relomp(" <<a<<", "<<b<<", "<<c<<", "<<d<<")="<<((1-gamma) * ratio(a,c) + gamma * ratio(b,d))<<"\n";
 	return ((1-gamma) * ratio(a,c) + gamma * ratio(b,d)) > 0.5;
 	//return (1-beta)*pow(a*ratio(b,a),alpa)+ (beta)*pow(b*ratio(a,b),alpa) < (1-beta)*pow(c*ratio(d,c),alpa)+ (beta)*pow(d*ratio(c,d),alpa);
